words-ranking/requirements.txt
quixstreams<2.7
python-dotenv